Yet another attempt at files with newline
Fixes: https://github.com/sayanarijit/xplr/issues/515

This is a proof of concept pr fix suport for handling files with newline
characters in them, without breaking existing api.

This is done by representating '\n' with '\0' in file names.

For example, `a\nb` becomes `a\0b`.

Since it's not possible to assing strings with `\0` in them to bash
variables, we either use something like `xxd` to encode/decode strings
into hex, or implement xplr helpers such as `xplr --pipe-msg-in <msg>`,
`xplr --pipe-result-out` which will read and print `\0` delimited
inputs, or provide equivalent abastraction to help with bash scripts.

While this makes it possible to copy, move and delete files with newline
characters, it's not possible to create such files due to input buffer's
limitations. Should it be fixed? I'm against it.

Lua api remains same. Changes in bash api shouldn't impact existing
config.

Another interestin gattempt with breaking changes by @jmcantrell is -
https://github.com/sayanarijit/xplr/pull/516